Move CRL verification code into crl module
